Add ChainDropdown Component for Chain Selection and Integrate into Navbar
Related to #753
